backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.92.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pydantic==1.10.5
python-dotenv==1.0.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
